dodoR-package {dodoR}	R Documentation

R functions I used for my master's thesis

Description

This is a personal package I created in order to automate the generation of reports, charts and CSV files concerning the use of Item Response Theory in a binary matrix of responses.

Details

Package: 	 dodoR

Type: 	 Package

Version: 	 1.0

Date: 	 2015-03-30

License: 	 What license is it under?

Author(s)

Douglas De Rizzo Meneghetti

Maintainer: Douglas De Rizzo Meneghetti <douglasrizzo@fei.edu.br>

See Also

mirt

[Package dodoR version 1.0 Index]