irt.report {dodoR}	R Documentation

Generate full IRT report based on mirt package

Description

This function generates a full report of an IRT run of the mirt package. It plots all possible charts for the test and each of the items; saves item parameters and item fit statistics (like Zhs and X2) into csv files; and can alternatively be used in conjunction with Sweave to generate a PDF report, since it prints the summaries of all relevant variables during a run.

Usage

irt.report(answers, out, out.stats = NULL, out.itemplots = NULL,
  out.testplots = NULL, out.densities = NULL, keys = NULL,
  itemtype = "3PL", method = "EM", optimizer = "NR", verbose = F,
  test_score = T, test_info = T, test_SE = T, test_infoSE = T,
  trace = T, info = T, se = T, score = T, infoSE = T, infotrace = T)

Arguments

answers
A data.frame containing examinees answers to items. Rows must represent examinees and columns represent items. If the data.frame is in binary format, keys is not required, but distractor analysis is not perfoermed.

out
A directory to which all files will be saved. If the directory does not exist, it will be recursively created.

out.stats
The directory to which CSV files will be saved. If the directory does not exist, it will be recursively created. if out.stats = NULL, it will be a subfolder of out.

out.itemplots
The directory to which item plots will be saved. If the directory does not exist, it will be recursively created. if out.itemplots = NULL, it will be a subfolder of out.

out.testplots
The directory to which test plots will be saved. If the directory does not exist, it will be recursively created. if out.testplots = NULL, it will be a subfolder of out.

out.densities
The directory to which parameter density plots will be saved. If the directory does not exist, it will be recursively created. if out.densities = NULL, it will be a subfolder of out.

keys
A vector containing the keys of the test, that is, the right answers for each item in the correct order.

itemtype
Item types to be used during parameter estimation. This parameter is passed directly to mirt. Default is '3PL', the three-parameter logistic model. For more options, consult the mirt manual.

method
Estimation method. This parameter is passed directly to mirt. Default is 'EM', the Expectation-Maximization algorithm. For more options, consult the mirt manual.

optimizer
Numerical approximation method used for optimizing convergence of estimation. This parameter is passed directly to mirt. Default is 'NR', the Newton-Raphson method. For more options, consult the mirt manual.

verbose
Defines how much output the algorithm must give. This parameter is passed directly to mirt. Default is FALSE. For more options, consult the mirt manual.

test_score
Whether or not test score plot should be generated.

test_info
Whether or not test information plot should be generated.

test_SE
Whether or not test standard error plot should be generated.

test_infoSE
Whether or not a plot containing both test information and standard error should be generated.

trace
Whether or not item trace plots should be generated.

info
Whether or not item information plots should be generated.

se
Whether or not item standard error plots should be generated.

score
Whether or not item score plots should be generated.

infoSE
Whether or not a plot containing both item information and standard error should be generated.

infotrace
Whether or not a plot containing both item information and trace should be generated.

Value

function returns an object of class SingleGroupClass (SingleGroupClass-class), the same one returned by the 'mirt-package' function 'mirt', in case additional processing is desired beyond the generation of plots and CSV files with parameters

Author(s)

Douglas De Rizzo Meneghetti

[Package dodoR version 1.0 Index]