normalize.responses {dodoR}	R Documentation

Normalize Range of Responses in a Matrix

Description

normalize.responses takes a response matrix and its keys and puts its values between [1; n of keys for current item]

Usage

normalize.responses(x, y)

Arguments

x
A response matrix.

y
A list containing the item keys.

Details

Although the use of a non-normalized response-matrix does not interfere in the estimation of IRT parameters, sometimes it is desirable to have item alternatives and its corresponding keys in a normalized range, especially when performing distractor analysis with irt.report.

So, for example, a list composed of the values [4, 3, 7] would become [2, 1, 3].

Value

A list containing the normalized response matrix in the first slot and its corresponding normalized keys in the second slot.

Author(s)

Douglas De Rizzo Meneghetti

[Package dodoR version 1.0 Index]